Add files via upload
GUI application that implements the classes that will be used to keep stock for a small 
corner store that sells only confectionery and reading material. For each item you will track two things: what it is, and how many items you have in stock. The derived classes will add one 
appropriate data member of their own.When a user chooses to add an item from the menu, use the Factory Method design pattern to 
return a pointer to either a Confectionery or Reading Material object from the data entered 
by the user. The factory method should take a Q String as an argument indicating what type of 
item is required, as well as the data used to construct the specific object. Note that the factory 
method should have only one createStock() function;